feat: add ratings with couses in gat courdede

+++ database/migrations/2025_09_19_185225_create_ratings_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
     {
         Schema::create('ratings', function (Blueprint $table) {
             $table->id();
-            $table->tinyInteger('rating');
+            $table->double('rating');
             $table->text('comment')->nullable();
             $table->foreignId('user_id')->constrained('users')->cascadeOnDelete();
             $table->foreignId('course_id')->nullable()->constrained('courses')->nullOnDelete();
